+++ src/wfg/wrap_ui/ui/systems/BaseSystem.java
@@ -3,7 +3,6 @@ package wfg.wrap_ui.ui.systems;
 import java.util.List;
 
 import com.fs.starfarer.api.input.InputEventAPI;
-import com.fs.starfarer.api.ui.UIPanelAPI;
 
 import wfg.wrap_ui.ui.panels.CustomPanel;
 import wfg.wrap_ui.ui.plugins.CustomPanelPlugin;
@@ -28,7 +27,7 @@ import wfg.wrap_ui.ui.plugins.CustomPanelPlugin.InputSnapshot;
  */
 public abstract class BaseSystem<
     PluginType extends CustomPanelPlugin<PanelType, PluginType>,
-    PanelType extends CustomPanel<PluginType, PanelType, ? extends UIPanelAPI>
+    PanelType extends CustomPanel<PluginType, PanelType>
 > {
 
     private PluginType plugin;