+++ changelog.md
@@ -1,6 +1,9 @@
 v0.2.1-beta
 - Misc. bug fixes
-- Made SpritePanel and SpritePanelWithTp a lot leaner in termns of composition
+- Made SpritePanel and SpritePanelWithTp a lot leaner in terms of composition
+- Added createTooltip and addTooltip to ComponentFactory for CustomPanelAPI independent Tooltips.
+- Added addCaptionValueBlock to ComponentFactory to have values with titles.
+- Removed ParentType from CustomPanel, as it does not need to be a CustomPanelAPI instance anymore.
 
 v0.2.0-beta
 - Created a carbon copy of DialogPanel because the API for vanilla version was too restrictive